lokabhasha-backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Vyoman Jain <75789232+VyoJ@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
sqlalchemy = "^2.0.35"
pydantic = "^2.9.2"
uvicorn = "^0.31.0"
pymysql = "^1.1.1"
mysql-connector-python = "^9.0.0"
python-dotenv = "^1.0.1"
cryptography = "^43.0.3"
passlib = "^1.7.4"
bcrypt = "^4.2.0"
pyjwt = "^2.9.0"
pytest = "^8.3.3"
httpx = "^0.27.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
